Changed to use OpenZeppelin SafeMath and StandardToken contracts and eliminate throw().

+++ contracts/FractionalERC20.sol
@@ -13,6 +13,6 @@ import "zeppelin/contracts/token/ERC20.sol";
  */
 contract FractionalERC20 is ERC20 {
 
-  uint public decimals;
+  uint8 public decimals;
 
 }
